Death and Attack State, possessing NPC and Model randomizer
Attack State: Now players can attack NPCs and other players from behind, killing them
AI and player Death State: If an NPC dies, the player that killed them has to extinguish the body before continuing with other tasks, if a player dies they get respawned in the body of another NPC.
Possessing NPC: When a player dies they take the place of an NPC in the world, it is already working, but we need to know if it's ok for you guys that when they do that there will be fewer NPCs in the area, I think this could be a nice feature for the game though, ratio players x NPC increase as players die.
Model Randomizer: We now spawn NPCs and Players with random models, but we need a few adjustments on the models for it to work better: Centralize the pivot point of the model; Rotate the character 90 degrees on the Y-Axis (Z-Axis in Blender); decrease the size of the characters to around 30% of its original size.